Create README.md
About Bag-It
A bag helps a person keep their items safe and intact, as well as completes a person’s look. Bag-It bag store sells diverse bags to both males and females in Ghana.
They stock a rich collection from various well-known brands all over the world. The goal of the Bag-It website is to make handbags, luggage, backpacks, purses and other types of bags from various these brands, more accessible to customers in Ghana. Currently, the e-commerce bag store is only available in Ghana.

Value Proposition
Value Proposition refers to the value that your company is trying to provide for its customers; that is, why your company exists.
Bag-It provides an online shopping website that offers and delivers all kinds of bags from the best brands worldwide to people all over Ghana. The brands are mostly found in European countries, however, Bag-It delivers the bags directly to the recipients in Ghana, in their comfort, as they are available to at the click of a button, and at cheaper prices

Revenue Stream Model
The revenues will be obtained by Bag-It online bag store through Profit Margin. Our primary source of revenue would be the low-profit margin on the sale of the bags.

System Design
High-Level Systems Overview and Architecture
The Bag-It e-commerce application was built with the Three-tiered Architecture and the Model- View-Controller (MVC) Framework.
Three-Tier Architecture
It is a client-server architecture in which the functional process logic, data access, computer data storage and user interface are developed and maintained as independent modules on separate platforms. Below are the tiers;
• Presentation Tier- The presentation tier is the front-end layer in the 3-tier system and consists
of the user interface. This layer is the graphical representation of the site, and can be described as the user interface, often accessible through a web browser and displays content and information useful to an end user. This tier will be built on the web technologies of HTML5, JavaScript, CSS, and communicates with other layers through API calls.
• Application/ Functional Tier-The application tier implements the logic behind the functionalities on the site. It the functional business logic which drives an application’s core capabilities. It will be written in PHP and jQuery.
• Data Tier- The data tier comprises of the database/data storage system. It represents the data being used by the site and provides an interface for the database being implemented. It will be written in MySQL.

Building the System
Key models/ layers for Bag-It
• Customer Interface
- Sign Up, Sign in and Sign Out
- Search feature
- Browse through catalog by categories and brands.
- Cart management
- Purchasing bags
• Admin Interface
For inventory purposes, the admin is also a customer, as she can buy bags as and when she desires. Aside the customer interface tasks, an admin can;
- Login and Sign Out
- Manage Database
- Add, Update, delete brands
- Add, Update, delete new categories
- Add, Update, delete new products

Technologies Used
Implementation
For the project, we used various programming languages, including;
- HTML – HTML is the abbreviated form of Hypertext Markup Language. It is a programming
language used in the creation of documents that are displayed in web browsers. HTML is one of the languages the team used in developing the front-end of the platform. On the Bag-it website, all pages which you see written content was made using HTML. For instance, the navigation bar was done using HTML. I did regex in HTML, using the pattern attribute.
- CSS – CSS stands for Cascading Style Sheet. CSS is the programming language utilized in designing HTML documents. CSS enables the use of fonts, colours and layouts on web pages. It was also used in developing the front-end. Both inline and external CSS was utilized for this project. It was used in designing the login and registration pages, the shop page and all the pages with exceptional designs. How the content of the pages is styled was done using CSS. The color, the font type, the hover of the navigation bar was styled using CSS.
- JavaScript – JavaScript handles the interactivity of the web pages. It is a front-end programming language. Alerts, button responses and form validation were made possible through JavaScript. I used JavaScript for the alerts, for instance, when you update quantity in cart, it will echo, “Cart Quantity updated”.
- PHP – Hypertext Preprocessor is the full meaning of PHP. It is a server-side programming language. It is utilized in backend development. PHP is responsible for interaction with the database; mainly inputting and retrieving data. The user being able to login, add to cart, etc. was done using PHP. PHP allows users to interact with the front-end while receiving information from the database.
14
- SQL – SQL is the abbreviated form of Structured Query Language. SQL was utilized in managing data in the relational database management system. Creating and populating the database was done with SQL. In creating the database, 9 tables were needed to store all the information that will be on the project.


Frameworks
- Bootstrap – Bootstrap is an open-source front-end framework. It contains CSS, JavaScript and HTML based templates for responsive web development. Bootstrap helped us to get access to components that helped in creating elements like the navbar, the carousel, login and registration cards and most of the admin work, products display page, etc.

Tools
- PhpMyAdmin – It is an open-source administration tool used for MySQL. It was used to easily create the database for the project.
- GitHub – GitHub is an open-source code hosting platform, which I used to store all my codes. This was for in case something happened to my laptop.
- Sublime Text – It is also an integrated development environment that was used to edit the codes for the project.
Application Programming Interface (API)
- PayPal API – Paypal is an online payment system. Its API allows for the system to be integrated into various web applications to handle payments. The PayPal JavaScript SDK was used to integrate payment to allow users to make the payment on products a user wants to purchase. We used a sandbox account to make and receive payments, thus money sent or received aren’t real.
- Paystack API – Paystack is another online payment system. Its API also allows the Bag-It ecommerce website to accept payment of its bags. We used the tests APIs; thus, all payments aren’t live.